Adicionando Exceptions

+++ spring/firtsTest/first/src/main/java/com/everton/first/service/ProdutoService.java
@@ -1,5 +1,6 @@
 package com.everton.first.service;
 
+import com.everton.first.exceptions.RecursoNaoEncontradoException;
 import com.everton.first.model.Produto;
 import com.everton.first.Repository.ProdutoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -19,8 +20,9 @@ public class ProdutoService {
         return produtoRepository.findAll();
     }
     //Buscar por id
-    public Optional<Produto> buscarPorId(Long id) {
-        return produtoRepository.findById(id);
+    public Produto buscarPorId(Long id) {
+        return produtoRepository.findById(id)
+                .orElseThrow(() -> new RecursoNaoEncontradoException("Produto não encontrado com ID: "+ id));
     }
 
     //Salvar produto
@@ -30,6 +32,10 @@ public class ProdutoService {
 
     //Deletar
     public void deletarProduto(Long id){
-        produtoRepository.deleteById(id);
+        if (!produtoRepository.existsById(id)){
+            throw new RecursoNaoEncontradoException("Produto não encontrado com ID: "+ id);
+        }else {
+            produtoRepository.deleteById(id);
+        }
     }
 }
